Add a placeholder widget for the minimap
This adds a placeholder image widget that will later be used to render
the minimap. It ties it's creation to the player blueprint's BeginPlay
event, creating the widget widget and adding it to the viewport.

The placeholder widget itself is just an image widget for the time
being, anchoring it's lower left corner to the lower left of the screen.
It has no contents as yet and is just a white box.