Modularise minimap
This restructures the minimap a bit, primarily to remove the requirement
to create it via the character's blueprint. Firstly, it moves the
minimap into its own folder to keep all assets related to it together
and make it easier to turn into a plugin later.

Secondly, it adds a SceneCapture2D blueprint that is added to the scene.
The BeginPlay() event on this blueprint takes over creating the minimap
widget and adding it to the viewport from the player, which will help
keep the minimap logic together in one place, to make it easier when we
try to turn this into a plugin later.

The idea of using a SceneCapture2D blueprint is that we may be able use
it to render the scene from a top down perspective to form the minimap
background, but that is not hooked up yet.

This also adds a render target and sets the widget background to use it.
Nothing is currently rendered to this target, but it's clear colour is
set to blue to confirm that it is indeed being used.