Draw scene on minimap
This hooks up the SceneCapture2D's CaptureComponent2D to the render
target, and since that is already hooked up to the minimap's Image
widget we now see the scene being rendered in the minimap.

It's set to output LDR (the render target is set to RGB8), and an
orthographic projection, but it's position is currently fixed up in the
sky and doesn't follow the player around.