Tie minimap camera position to player
This pins the minimap camera position above the player. It uses a
blueprint to get the player pawn's position, set the camera's X & Y to
match and add an offset to the Z coordinate (which arbitrarily defaults
to +100).

This gives us a little top down view in the corner, which is a minimap
of sorts.