Tie minimap rotation to the camera
This gets the main camera's rotation vector and uses it's yaw value to
set the minimap camera's rotation, which results in the minimap's
up/north direction always matching the player's camera.

The reason for using the camera's rotation and not the player pawn's
rotation is because in the third person style the two are detached and
it's a little disorienting if the map is aligned with the player rather
than the camera.