Add variable to adjust minimap center
This adds a variable to allow the center of the minimap to be offset -
if it is set to 0 than the center of the minimap will line up with the
player, but if it is set to something like 512 (and the minimap render
target size is 2048) than the player would be 1/4 up the minimap to
allow the map to show more of what is in front of the player than
behind.

This is added as a variable to allow the it to be set as appropriate on
a case by case basis.