Add minimap widget C++ class
This reparents the minimap widget blueprint on a C++ class with a
blueprint implementable function to return the panel area to C++ code
that will be used to display the icons on the minimap.

This allows the minimap UI to be customised to suit the game while still
allowing the C++ code to draw the icons in the right positions. To
demonstrate this, a simple black border is added around the minimap.