Move minimap widget creation into Minimap blueprint
This moves the minimap widget creation out of the ScreenCapture2D
blueprint and into a blueprint derived from the C++ minimap class,
because structurally we want the minimap class to be the logical top
level of the minimap, and the ScreenCapture2D will be one component
responsible for the background, which therefore should not be the
component responsible for adding the widget to the viewport.

We do still use a blueprint to add the widget to the viewport, as this
allows for more flexibility if a developer wanted to place it somewhere
other than the viewport, such as on an in-game GPS screen.

This also adds a property to tell the minimap C++ class where to find
the minimap UI, and gets references to the widget and panel that will
eventually hold the icons from the C++ code.