Draw test icon on minimap
This finishes passing a reference to the minimap widget to the C++ code
via assigning it in the blueprint that creates it and adds it to the
viewport (or wherever the game's UI designer would prefer). The widget
blueprint implements a function that allows the C++ code to get the
specific panel that is to hold the icons (and clip them out of bounds)
and creates an empty test icon to prove that it is working.