Add actor component to set minimap icon on individual actors
This adds an ActorComponent that can be used to set the minimap icon for
a specific actor, rather than all actors of a given class (the two can
also be used in conjunction, with the class icon used to set a default
that can be overridden on specific actors).

Since the test actors I'm using are all instances of StaticMeshActor
this gives me a way to set icons appropriate for the specific mesh each
actor is using. The player icon is still set via the minimap class
legend.