Move scene capture into minimap class, export & synchronise properties
This replaces the SceneCapture2D blueprint that we have been using to
provide the minimap background with an equivalent component in the
minimap C++ class, with the relevant default properties set in the
constructor.

Properties for the minimap scale, camera height and offset for the
center are exported through the minimap component, and these can be
adjusted on demand.

To test these out, more key bindings have been added to the level
blueprint so holding the arrow keys will move the minimap center, and
the numpad +/- keys will zoom in and out.

I've left the original SceneCapture2D blueprint under
"Minimap/ObsoletePrototype" - although it's not being used anymore it's
still nice to see how quickly and easily the basic functionality could
be created in a blueprint.