Only show static actors on background
This cuts down on the number of actors drawn on the background, by
enabling the show only list and populating it with only the actors that
have static root components.

In practical terms this eliminates drawing actors we have an icon for,
but it is also good for performance since the background will have far
less to render. In addition, the scene capture component is set to only
capture when it moves so that it won't have to update every frame if the
player is stationary.