Reduce workload each tick - cache values & cull icons
This works to reduce the amount of work the minimap does each tick by
two measures - firstly, the intermediate results of calculations based
on properties that aren't expected to change much at runtime are cached
so that we only need to calculate them on demand, such as when the panel
changes size or when the minimap scale or centre is updated at runtime.
For the former case we detect when the panel's cached geometry changes
and update our calculations, and in the later case we require that
whoever updated the property let us know by calling
UpdateCachedProperties (and the level blueprint is updated to do this).

Secondly, we calculate the maximum distance from the player than an icon
might possibly be drawn on the minimap and cull any that are out of
range without even calculating their rotation. The maximum distance is
calculated only when cached values are updated, and the check for each
object is as small and light as possible avoiding the need do a full
distance calculation for each object. As a result it won't cull things
very tightly, but it will quickly cull things that are way too far out
of range and the panel can cull any remaining objects that are slightly
out of range.