Allow minimap to follow actors other than player & override rotation
This adds a follow_actor property to the minimap, which can be set to
have the minimap follow an actor other than the player (e.g. it could
follow the second player in split screen, an NPC, whatever).

This also exposes a boolean to choose whether the minimap's rotation is
locked to the camera or not, and exposes the yaw as a property to
manually set the rotation as desired.

To test the manual rotation, the level blueprint is extended so that
numpad 7/9 will unlock the minimap rotation from the camera and rotate
it. To test following other actors numpad 1/3 will set the minimap to
follow any spheres that have been spawned with F.

Numpad 5 will reset the state - set the minimap to follow the player,
lock the rotation to the camera and center the player in the middle of
the minimap.