Add ways to override which actors appear on background
By default the minimap will add all static actors to the background
layer, but there are cases where the dev might want something else -
e.g. they might not want a roof or other complex static geometry
included, or they might have something else in mind for the background
entirely.

This adds a checkbox in the minimap properties that can be used to
disable automatically adding static actors to the background, and
another checkbox in the minimap components that can be attached to
individual actors to indicate if they should or should not be included
on the background.