Fix issues with non-square minimaps
This fixes an issue where the background would be stretched and not line
up with the icons if the minimap panel was rectangular. It compensates
for the stretching by setting the render scale on the background widget,
properly adjusts the culling distance and adjusts the offset used to
capture the background image to line up with the player's icon.

As an added bonus, more of the camera rotation calculation is cached
now.